Post admin view, Post controller
Post Admin View:
customized the "managed posts" screen

Post Controller:
actionDelete() now only deletes on POST requests, 400 error generated
otherwise
Added afterDelete() to clear comments, and update tags.

+++ protected/controllers/PostController.php
@@ -108,11 +108,15 @@ class PostController extends Controller
 	 */
 	public function actionDelete($id)
 	{
-		$this->loadModel($id)->delete();
+		if(Yii::app()->request->isPostRequest){
+			$this->loadModel($id)->delete();
 
-		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-		if(!isset($_GET['ajax']))
-			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+			if(!isset($_GET['ajax'])){
+				$this->redirect(array('index'));
+			}
+		}else{
+			throw new CHttpException(400, 'Invalid request.  Please do not repeat this request again.');
+		}
 	}
 
 	/**
